+++ CHANGELOG.md
@@ -10,7 +10,8 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `URLSearchParams::when` conditional method to ease component building logic.
 - `Modifier::prependQueryParameters` returns a modifier with prepend query paramters
 - `Modifier::when` conditional method to ease component building logic.
-- `Modifier::normalizeHostIp` returns the host as normalized by the WHATWG algorithm
+- `Modifier::normalizeIp` returns the host normalized for IPv6 and IPv4 addresses
+- `Modifier::normalizeHost` returns the host as normalized by the WHATWG algorithm
 - `Modifier::with*` method from the underlying `Uri` object are proxy to improve DX.
 - `Query::decoded` the string representation of the component decoded.
 - `URLSearchParams::decoded` the string representation of the component decoded.